# Improve `include_headers` script