### What does this PR do?

This PR improves the `include_headers` script in 3 ways by:
- using Buffered IO when dumping included headers. Currently, the script does a lot of calls to write.
- caching the path search result, since the result is the same for each `(srcPath, headerName, includeDirs)`
- skipping the allocation when appending the current folder to the search tree

Benchmark:
```sh
>> time go generate -mod=mod -tags linux_bpf ./pkg/security/ebpf/compile.go
```

Before:
```
Executed in    1.89 secs    fish           external
   usr time    0.53 secs  234.00 micros    0.53 secs
   sys time    1.29 secs  221.00 micros    1.29 secs
```

After:
```
Executed in  727.89 millis    fish           external
   usr time  609.56 millis  251.00 micros  609.31 millis
   sys time  593.48 millis  237.00 micros  593.24 millis
```

### Motivation

Speed-up the `include_headers.go` tool. This is run in dev environment and on CI so no user impact but still nice to have.

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.